Добавлено поле hours в сущность LessonTemplate. Контроллер разделен на несколько, для облегчения расшения
