+++ backend/src/main/java/com/demonorium/database/entity/LessonTemplate.java
@@ -56,6 +56,11 @@ public class LessonTemplate implements PartOfSource {
     private String note;
 
     /**
+     * Количество часов, отведённое на предмет
+     */
+    private int hours;
+
+    /**
      * Список преподавателей
      */
     @ManyToMany(cascade = CascadeType.REFRESH, fetch = FetchType.EAGER)
@@ -68,9 +73,10 @@ public class LessonTemplate implements PartOfSource {
     @OneToMany(mappedBy = "template", cascade = CascadeType.REMOVE, fetch = FetchType.EAGER)
     private Set<Lesson> lessons = new HashSet<>();
 
-    public LessonTemplate(String name, String note, Source source) {
+    public LessonTemplate(String name, String note, int hours, Source source) {
         this.name = name;
         this.note = note;
         this.source = source;
+        this.hours = hours;
     }
 }
