+++ frontend/src/components/screens/edit-source/LessonTemplateEditor.tsx
@@ -1,15 +1,22 @@
 import * as React from 'react';
 import {useState} from 'react';
 import ItemListEditor from "./ItemListEditor";
-import {addPlace, changePlace, removePlace, SourcesRepresentation} from "../../../store/sourceMap";
-import {Place} from "../../../database";
+import {
+    addPlace,
+    addTemplate,
+    changePlace,
+    changeTemplate,
+    removePlace, removeTemplate,
+    SourcesRepresentation
+} from "../../../store/sourceMap";
+import {LessonTemplateDto, Place} from "../../../database";
 import axios from "axios";
 import {useAppDispatch, useAppSelector} from "../../../store/hooks";
 import {Editor} from "../../modals/ModalEditor";
 import {Grid, ListItemText, TextField, Typography} from "@mui/material";
 
 
-interface PlaceListEditorProps {
+interface LessonTemplateEditorProps {
     isSelect: boolean;
     overrideTitle?: string;
 
@@ -17,47 +24,43 @@ interface PlaceListEditorProps {
     /**
      * Закрыть это окно
      */
-    requestClose?: (item: Place) => {};
+    requestClose?: (item: LessonTemplateDto) => {};
 }
 
-export default function PlaceListEditor(props: PlaceListEditorProps) {
+export default function LessonTemplateEditor(props: LessonTemplateEditorProps) {
     const user = useAppSelector(state => state.user);
     const dispatch = useAppDispatch();
 
     const defaultState = {
         id: -1,
-        source: props.source.source.id,
-        building: "",
-        auditory: ""
+        name: "",
+        hours: 0,
+        defaultTeachers: new Array<number>()
     }
-    const [state, setState] = useState<Place>(defaultState);
-    const handleChange = (prop: keyof Place) => (event: React.ChangeEvent<HTMLInputElement>) => {
+    const [state, setState] = useState<LessonTemplateDto>(defaultState);
+
+    const handleChange = (prop: keyof LessonTemplateDto) => (event: React.ChangeEvent<HTMLInputElement>) => {
         let value = event.target.value;
-        if (value.length > 8) {
-            if (prop == "building" || prop == "auditory")
-                value = value.substr(0, 8);
-            else if (value.length > 50)
-                value = value.substr(0, 50);
-        }
+        if (value.length > 50)
+            value = value.substr(0, 50);
+
         setState({ ...state,
             [prop]: value
         });
     };
 
-    const editor: Editor<Place> = {
+    const editor: Editor<LessonTemplateDto> = {
         onPartCreate: (item) => {
-            axios.get("api/create/place", {
+            axios.get("api/create/lessonTemplate", {
                 auth: user,
                 params: {
                     sourceId: props.source.source.id,
-
+                    name: item.name,
                     note: item.note,
-                    auditory: item.auditory,
-                    building: item.building
                 }
             }).then((response) => {
                 item.id = response.data;
-                dispatch(addPlace({item: item, source: props.source.source.id}))
+                dispatch(addTemplate({item: item, source: props.source.source.id}))
             });
         },
         onPartUpdate: (item) => {
@@ -65,7 +68,7 @@ export default function PlaceListEditor(props: PlaceListEditorProps) {
                 auth: user,
                 params: item
             }).then((response) => {
-                dispatch(changePlace({item: item, source: props.source.source.id}))
+                dispatch(changeTemplate({item: item, source: props.source.source.id}))
             });
         },
 
@@ -81,7 +84,7 @@ export default function PlaceListEditor(props: PlaceListEditorProps) {
             }
             return false;
         },
-        changeItem(item: Place | undefined): void {
+        changeItem(item): void {
             if (item == undefined) {
                 setState(defaultState);
             } else {
@@ -91,40 +94,34 @@ export default function PlaceListEditor(props: PlaceListEditorProps) {
 
 
         UI: (
-         <Grid container spacing={2}>
-             <Grid item xs={4}>
-                 <Typography variant="h6">Место: </Typography>
-             </Grid>
-             <Grid item xs={4}>
-                 <TextField fullWidth label="Аудитория" defaultValue={state.auditory} onChange={handleChange("auditory")}/>
-             </Grid>
-             <Grid item xs={4}>
-                 <TextField fullWidth label="Здание / Корпус" defaultValue={state.building} onChange={handleChange("building")}/>
-             </Grid>
-             <Grid item xs={12}>
-                 <TextField fullWidth label="Заметка" defaultValue={state.note} onChange={handleChange("note")}/>
-             </Grid>
-         </Grid>
+            <Grid container spacing={2}>
+                <Grid item xs={12}>
+                    <TextField fullWidth label="Название" defaultValue={state.name} onChange={handleChange("name")}/>
+                </Grid>
+                <Grid item xs={12}>
+                    <TextField fullWidth label="Заметка" defaultValue={state.note} onChange={handleChange("note")}/>
+                </Grid>
+            </Grid>
         )
     }
 
-    return <ItemListEditor<Place>
+    return <ItemListEditor<LessonTemplateDto>
         requestClose={props.requestClose}
         listTitle={props.overrideTitle? props.overrideTitle : "Список мест проведения занятий"}
-        list={props.source.places}
+        list={props.source.templates}
         isSelect={props.isSelect}
         constructor={(item, index) =>
-            <ListItemText primary={item.auditory + " / " + item.building} secondary={item.note} />
+            <ListItemText primary={item.name} secondary={item.note} />
         }
         remove={(item) => {
             axios.get("api/delete/place", {
                 auth: user,
                 params: {id: item.id}
             }).then(() => {
-                dispatch(removePlace({item: item, source: props.source.source.id}))
+                dispatch(removeTemplate({item: item, source: props.source.source.id}))
             });
         }}
         editorTitle="Место проведения занятия"
         editor={editor}
-         />;
+    />;
 }
